+++ tests/distributions/test_univariate.py
@@ -3,7 +3,7 @@ import pytest
 import tensorflow as tf
 
 from tfsnippet.distributions import *
-from tfsnippet.utils import scoped_set_check_numerics
+from tfsnippet.utils import scoped_set_config, settings
 
 
 class NormalTestCase(tf.test.TestCase):
@@ -27,7 +27,7 @@ class NormalTestCase(tf.test.TestCase):
             np.testing.assert_allclose(normal.logstd.eval(), logstd)
 
     def test_check_numerics(self):
-        with scoped_set_check_numerics(True):
+        with scoped_set_config(settings, check_numerics=True):
             normal = Normal(mean=0., std=-1.)
             with self.test_session():
                 with pytest.raises(
@@ -85,7 +85,7 @@ class UniformTestCase(tf.test.TestCase):
             self.assertEqual(uniform.maxval.eval(), 2.)
 
     def test_check_numerics(self):
-        with scoped_set_check_numerics(True):
+        with scoped_set_config(settings, check_numerics=True):
             uniform = Uniform(minval=-1e100, maxval=1e100)
             with self.test_session():
                 with pytest.raises(