+++ tests/test_config.py
@@ -1,0 +1,10 @@
+import unittest
+
+import tfsnippet as spt
+
+
+class TFSnippetConfigTestCase(unittest.TestCase):
+
+    def test_tfsnippet_config(self):
+        self.assertTrue(spt.settings.enable_assertions)
+        self.assertFalse(spt.settings.check_numerics)