+++ tfsnippet/shortcuts.py
@@ -4,6 +4,7 @@ This package provides shortcuts to utilities from second-level packages.
 
 from .dataflows import *
 from .utils.reuse import *
+from .utils.config_ import *
 from .utils.config_utils import *
 
 __all__ = [
@@ -14,7 +15,10 @@ __all__ = [
     'get_reuse_stack_top', 'instance_reuse', 'global_reuse',
     'VarScopeObject',
 
-    # from tfsnippet.utils.config
+    # from tfsnippet.utils.config_
+    'settings',
+
+    # from tfsnippet.utils.config_utils
     'Config', 'ConfigField',
-    'config_defaults', 'register_config_arguments',
+    'get_config_defaults', 'register_config_arguments',
 ]